Refactor interactive flow
The interactive flow has been added in a separate folder so that every interactive flow can have its own file.
