+++ lib/api/create-submenu-patch.js
@@ -61,7 +61,7 @@ export default async function createMenuPatch(menu, globs, options = {}) {
 	if (options.save) {
 		let buffer = dbpf.toBuffer();
 		let { output = 'Submenu patch.dat' } = options;
-		let outputPath = path.resolve(process.cwd(), output);
+		let outputPath = path.resolve(directory, output);
 		fs.writeFileSync(outputPath, buffer);
 		logger.info(chalk.green(`Saved to ${outputPath}`));
 	}
