+++ lib/cli/flows/index.js
@@ -1,0 +1,1 @@
+export { default as submenu } from './submenu.js';
