+++ lib/cli/flows/submenu.js
@@ -1,5 +1,58 @@
-// # constants.js
-export const Menu = {
+// # submenu.js
+import chalk from 'chalk';
+import path from 'node:path';
+import fs from 'node:fs';
+import * as prompts from '../prompts.js';
+
+// # submenu()
+// Contains the interactive flow for getting the options to pass to the submenu 
+// patch command.
+export default async function submenu() {
+	let dir = process.cwd();
+	let thisDir = await prompts.confirm({
+		message: `Are the files located in this directory? ${chalk.cyan(dir)}`,
+	});
+	if (!thisDir) {
+		dir = await prompts.fileSelector({
+			message: 'Select the directory where the files are located',
+			basePath: process.cwd(),
+			type: 'file+directory',
+			filter(info) {
+				if (info.isDirectory()) return true;
+				let ext = path.extname(info.path).toLowerCase();
+				return ['.sc4lot', '.sc4desc', '.sc4model', '.dat'].includes(ext);
+			},
+		});
+	}
+	let info = await fs.promises.stat(dir);
+	if (!info.isDirectory()) {
+		dir = path.dirname(dir);
+	}
+	let menu = await prompts.nestedList({
+		name: 'menu',
+		message: 'What menu do the items need to be added to ?',
+		type: 'list',
+		pageSize: 15,
+		choices: {
+			...Menu,
+			'Custom button id...': 'custom',
+		},
+	});
+	if (menu === 'custom') {
+		({ menu } = await prompts.hex({
+			name: 'menu',
+			message: 'Input a custom button id (e.g 0x5c43f355)',
+		}));
+		menu = +menu;
+	}
+	let output = await prompts.input({
+		message: 'Where do you want to save the patch?',
+		default: './submenu_patch.dat',
+	});
+	return [menu, '*.{dat,sc4*}', { output, directory: dir }];
+}
+
+const Menu = {
 	Residential: {
 		R$: 0x93DADFE9,
 		R$$: 0x984E5034,
