+++ lib/cli/setup-cli.js
@@ -7,12 +7,10 @@ import { program as commander, Command } from 'commander';
 import * as commands from './commands/index.js';
 import * as prompts from './prompts.js';
 import inquirer from 'inquirer';
-import { glob } from 'glob';
 import { DBPF, Savegame, FileType, ZoneType } from 'sc4/core';
 import * as api from 'sc4/api';
 import version from './version.js';
 import PipeManager from 'sc4/api/pipe-manager.js';
-import createMenuPatch from 'sc4/api/create-submenu-patch.js';
 import { hex } from 'sc4/utils';
 
 const Style = {
